fix: add better empty states to project view/toolwindow
